Moved TokenPositionInfo.tsx to section/pools/components
