+++ src/i18n/locales/en/translations.json
@@ -13,7 +13,6 @@
   "toast.sidebar.referendums.aye": "Aye",
   "toast.sidebar.referendums.nay": "Nay",
   "toast.sidebar.referendums.value": "{{ value, compact }} HDX ({{ percent, bignumber(type: 'percentage') }}%)",
-  "tokenPositionInfo.tooltip": "Sum of underlying asset and LRNA:",
   "rpc.change.modal.title": "Change RPC",
   "rpc.change.modal.column.name": "Name",
   "rpc.change.modal.column.status": "Status",
@@ -354,6 +353,7 @@
   "wallet.assets.farmingPositions.header.shares": "Asset shares",
   "wallet.assets.farmingPositions.header.value": "Current value",
   "wallet.assets.farmingPositions.data.date": "{{ date, dd.MM.yyyy }}",
+  "pools.tokenPositionInfo.tooltip": "Sum of underlying asset and LRNA:",
   "farms.positions.header.title": "Farm positions",
   "farms.positions.joinedFarms.button.label": "Details",
   "farms.positions.join.button.label": "Join",
