+++ src/sections/pools/pool/positions/LiquidityPosition.tsx
@@ -27,7 +27,7 @@ import { DisplayValue } from "components/DisplayValue/DisplayValue"
 import { useDisplayPrice } from "utils/displayAsset"
 import { BN_0 } from "utils/constants"
 import Skeleton from "react-loading-skeleton"
-import { TokenPositionInfo } from "components/TokenPositionInfo/TokenPositionInfo"
+import { TokenPositionInfo } from "../../components/TokenPositionInfo"
 
 type Props = {
   pool: OmnipoolPool
